// src/app/models/entidades.enum.ts
export enum Entidades {
  ALUNO = "aluno",
  FACULDADE = "faculdade",
  PROFESSOR = "professor",
  ALUNO_FACULDADE = "aluno-faculdade",
  PROFESSOR_FACULDADE = "professor-faculdade",
  AVALIACAO_FACULDADE = "avaliacao-faculdade",
  COMENTARIO_FACULDADE = "comentario-faculdade",
  AVALIACAO_PROFESSOR = "avaliacao-professor",
  COMENTARIO_PROFESSOR = "comentario-professor",
  FORMACAO_PROFESSOR = "formacao-professor",
}